+++ RELEASE_NOTES.md
@@ -12,6 +12,7 @@ _Date TBD_
   * remove has open surveillance, has closed surveillance, has open nonconformities, has closed nonconformities
   * add surveillanceCount, openNonconformityCount, closedNonconformityCount fields
 * Cache results of /certified_products/pending to improve view time by ~683%
+* Add Surveillance Friendly ID to Surveillance (Basic) CSV file download
 * Add optional "fields" parameter to /certified_products call to allow only sending back a custom list of fields
 
 ---